Add method `Schema::parse_str_with_list`
Add a new method `Schema::parse_str_with_list` that can parse
a root schema along with a list of named schemas.  This method
can be used to parse a root union that has named schema references,
for example.

This new method combines the functionality of `Schema::parse_str`
with `Schema::parse_list`.