fix #226: Index-Out-Of-Bounds panic when using #[serde(skip_serializing_if=..)] (#227)
* fix #226: Index-Out-Of-Bounds panic when using #[serde(skip_serializing_if = "Option::is_none")]

* Issue #225 - Compare the deserialized result with the serialized one

Signed-off-by: Martin Tzvetanov Grigorov <mgrigorov@apache.org>

* Issue #226: Serialize the RecordField's default when skipping a field

We need to serialize something, otherwise the deserialization does not
know that something has been skipped and tries to read a value according
to the schema

Signed-off-by: Martin Tzvetanov Grigorov <mgrigorov@apache.org>

* fmt

Signed-off-by: Martin Tzvetanov Grigorov <mgrigorov@apache.org>

* Do not try to track the fields with item_count.

Serde's `skip` breaks it because it does not notify us when
`#[serde(skip_serialize)]` is used.

Drop the support for Struct tuple (e.g. `struct S(A, B, C)` - it is hard
to map it to Avro record schema. It should still work for Avro array.

Signed-off-by: Martin Tzvetanov Grigorov <mgrigorov@apache.org>

* clippy + fmt

Signed-off-by: Martin Tzvetanov Grigorov <mgrigorov@apache.org>

* Add ASLv2 header for the new IT test

Signed-off-by: Martin Tzvetanov Grigorov <mgrigorov@apache.org>

---------

Signed-off-by: Martin Tzvetanov Grigorov <mgrigorov@apache.org>
Co-authored-by: Elad Yosifon <me@elad-yosifon.xyz>
Co-authored-by: Martin Tzvetanov Grigorov <mgrigorov@apache.org>