refactor: Simplify the UUID decoding logic
In the original implementation we first read the length and then
read the remainder either as `Fixed` if the length is 16 and `String`
if it is not. It uses a temporary `Vec` to be able to backtrack if
`Fixed` is wrong.

The backtrack can only happen if it fails to read a `Fixed` or if
it gets a different type than `Fixed` from `decode_internal`. The
last case can never happen, as `decode_internal` always returns the
type corresponding to the schema. The first case can only happen if
there are not enough bytes, because any byte pattern is valid for
`Fixed`. If `Fixed` cannot read enough bytes, than `String` won't
be able to read enough bytes either.

I did not add a fallback from `Uuid::from_slice` to `Uuuid::parse_str`
as an UUID string is at least 32 characters, so it would fail anyway.