fix: documentation linked to private items (#327)
I've made the items public, by making the `util` module public. Any
items in there that should not be public have been marked `pub(crate)`.

As `set_serde_human_readable` and `max_allocation_bytes` are now also
available through the `util` path, I've deprecated them in the root of
the crate. As you can't deprecate a re-export, I've done this by wrapping
them in a function and deprecating that function.

I've also updated the documentation of `set_serde_human_readable` and
`max_allocation_bytes` to explain the return value and that it can be
different than the user expects.