feat!: resolve schema(ta) in the builder of `Writer` (#328)
This moves the resolving error to the place where it should happen
and simplifies `Writer` functions.

This is a breaking change, as the builder now returns a `AvroResult`.

The actual relevant code is in the first 400 lines of `avro/src/writer.rs`.
All other changes are adding `?` or `.unwrap()` to test code.