feat: Implement support for `#[serde(flatten)]` (#359)
* chore: Move all Serde related modules to the `serde` module

* feat: Implement support for `#[serde(flatten)]`

This is done by adding a `#[avro(flatten]` attribute so that the
schema (for that field) is also flattened, and by adding support
in `SchemaAwareWriteSerializer` for serializing a struct via Map
instead of Struct.

`flatten` does not work with `to_value`, as `to_value` does not have
access to the schema.

* fix: Handle duplicate fields in when flatten is used

---------

Co-authored-by: default <admin@kriskras99.nl>